requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
Django==3.2.4
django-appconf==1.0.4
django-compressor==2.4.1
django-environ==0.4.5
django-solo==1.1.5
django-tinymce==3.3.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
